refactor: rename blinking algorithm to attention level
